chore: move error directory into utils
