+++ components/study/signup/StudySignupForm.tsx
@@ -11,7 +11,6 @@ import { MdOutlineSignalCellularAlt } from 'react-icons/md'
 import { RiStackOverflowLine } from 'react-icons/ri'
 import { z } from 'zod'
 
-import { ServerErrorType } from '@/app/api/errors/types'
 import { HttpStatusCode } from '@/app/api/utils/httpConsts'
 import { ServerResponse } from '@/app/api/utils/response'
 import LoadingSpinner from '@/components/common/LoadingSpinner'
@@ -22,6 +21,7 @@ import { Input } from '@/components/ui/input'
 import { Label } from '@/components/ui/label'
 import { Textarea } from '@/components/ui/textarea'
 import { API_ENDPOINTS } from '@/constants/apiEndpoint'
+import { ServerErrorType } from '@/lib/errors/types'
 import { fetchData } from '@/lib/fetch'
 import { Study } from '@/types'
