+++ lib/errors/types.ts
